+++ psalmer/hymnal/catalog/__init__.py
@@ -1,0 +1,1 @@
+from .hymnal_lib import HymnalLib
