+++ psalmer/hymnal/finder/__init__.py
@@ -1,0 +1,3 @@
+from .hymn_finder      import HymnFinder
+from .file_hymn_finder import FileHymnFinder
+from .db_hymn_finder   import DbHymnFinder
