+++ psalmer/hymnal/meta/__init__.py
@@ -1,0 +1,3 @@
+from .hymnal_meta import HymnalMeta
+from .range_meta import RangeMeta
+from .hymn_meta import HymnMeta
