+++ psalmer/hymnal/meta/hymn_meta.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+@dataclass
+class HymnMeta:
+    hymnal_id: int
+    id       : int
+    fmt      : str
+    title    : str
