+++ psalmer/hymnal/meta/hymnal_meta.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+@dataclass
+class HymnalMeta:
+    id   : int
+    code : str
+    title: str
