+++ psalmer/hymnal/meta/range_meta.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+@dataclass
+class RangeMeta:
+    hymnal_id      : int
+    id             : int
+    starting_prefix: str
+    ending_prefix  : str
